Allow pushing large object over HTTP (again)
`client_body_max_size` is added by
<https://github.com/turnkeylinux-apps/gitea/pull/7>,
but it does not work for named location due to the nginx bug
<https://trac.nginx.org/nginx/ticket/2322>.
According to the ticket, at least nginx/1.20.2 and nginx/1.21.6 are
affected.

To workaround the issue, specify `client_body_max_size` for non-named
location `/`.

Tested with nginx/1.18.0 (turnkey-gitea-17.2-bullseye-amd64) and this
change fixes the issue as expected.